Create class Menu
